backend/requirements.txt
# Core Flask and general utilities (kept pins)
Flask==3.1.1
flask-cors==6.0.0
gunicorn==22.0.0
pypdf==5.6.0
firebase-admin==6.4.0 # Kept specific as it's an external library
requests==2.32.3 # Kept specific as it's an external library

# Other general Python libraries (kept pins for consistency if they're not causing conflicts)
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
docstring_parser==0.16
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3


# Google Cloud and related core libraries - REMOVED VERSION PINS
# Let pip determine the compatible versions to resolve protobuf conflicts.
google-api-core
google-auth
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-dlp
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-genai
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
proto-plus
protobuf